requirements.txt
numpy
pandas
scikit-learn
lightgbm
xgboost
torch
requests
aiohttp
openpyxl
pykrx
tqdm
python-dotenv
